Se crea nueva hook personalizado para usar en la cartapp para ser mas legible el codigo
